+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2019-08-17
+### Added
+- Sitemap controller.
+- Google Play Store link.
+- Video in project page.
+
 ## [0.0.8] - 2019-08-16
 ### Added
 - Pages of categories and topics.
@@ -54,7 +60,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - First deployment version of this project.
 - This CHANGELOG file even this is not a open source project.
 
-[Unreleased]: https://github.com/ccmikechen/Birdiy-server/compare/v0.0.8...HEAD
+[Unreleased]: https://github.com/ccmikechen/Birdiy-server/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/ccmikechen/Birdiy-server/releases/tag/v0.1.0
 [0.0.8]: https://github.com/ccmikechen/Birdiy-server/releases/tag/v0.0.8
 [0.0.7]: https://github.com/ccmikechen/Birdiy-server/releases/tag/v0.0.7
 [0.0.6]: https://github.com/ccmikechen/Birdiy-server/releases/tag/v0.0.6